version 2.0 for lecture 7
