+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.20.0)
-project(lecture7 VERSION 0.1.0)
+project(lecture8 VERSION 0.1.0)
 
 
 # Compilation flags
@@ -7,33 +7,19 @@ add_compile_options(-Wall)
 
 include_directories(lecture7/include)
 # Add executable
-add_executable(lecture7 lecture7/src/main.cpp)
+add_executable(lecture7 lecture7/src/main.cpp lecture7/src/date.cpp)
 
-
-# add_custom_command(TARGET lecture3 POST_BUILD
-#     COMMAND ./lecture3 | c++filt -t
+# add_custom_command(TARGET lecture8 POST_BUILD
+#     COMMAND valgrind --show-error-list=yes --leak-check=full --show-leak-kinds=all --track-origins=yes ./lecture8
 #     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-#     COMMENT "Resolve name mangling"
-#     VERBATIM
-# )
-
-# add_custom_target(demangle ALL
-#     COMMAND ./lecture3 | c++filt -t
-#     DEPENDS lecture3
-#     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-# )
-
-# add_custom_command(TARGET lecture3 POST_BUILD
-#     COMMAND valgrind --show-error-list=yes --leak-check=full --show-leak-kinds=all --track-origins=yes ./lecture3
-#     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-#     COMMENT "Running Valgrind after building lecture3"
+#     COMMENT "Running Valgrind after building lecture8"
 #     VERBATIM
 # )
 
 
 # add_custom_target(run_valgrind ALL
-#     COMMAND valgrind --show-error-list=yes --leak-check=full --show-leak-kinds=all --track-origins=yes ./lecture3
-#     DEPENDS lecture3
+#     COMMAND valgrind --show-error-list=yes --leak-check=full --show-leak-kinds=all --track-origins=yes ./lecture8
+#     DEPENDS lecture8
 #     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
 # )
