+++ lecture7/include/code10.hpp
@@ -1,0 +1,5 @@
+class MyClass {
+   public:
+    static int count;  // Declaration of static attribute
+    
+};
