--- lecture7/include/code13.hpp
@@ -1,8 +1,0 @@
-class MathUtils {
-   public:
-    static int add(int a, int b) {
-        return a + b;
-    }
-};
-
-int result{MathUtils::add(3, 4)};  // No need to create a MathUtils object
