+++ lecture7/include/code17.hpp
@@ -1,0 +1,8 @@
+class Counter {
+   private:
+    static int count;
+
+   public:
+    static void increment() { count++; }
+    static int get_count() { return count; }
+};
