+++ lecture7/include/code18.hpp
@@ -1,0 +1,14 @@
+#include<iostream>
+
+class MyClass {
+   public:
+    // Constructor
+    MyClass() {
+        std::cout << "Object created!\n";
+    }
+
+    // Destructor
+    ~MyClass() {
+        std::cout << "Object destroyed!\n";
+    }
+};
