+++ lecture8/include/code3.hpp
@@ -3,24 +3,58 @@
 
 class Base {
    public:
-    Base(int base_value = 50) : base_{base_value} {
-        std::cout << "----- Base ctor"<< '\n';
-        std::cout << "base_: " << base_ << '\n';
+    Base() : base_pub_{10},
+             base_pro_{20},
+             base_pri_{30} {
+        std::cout << "Initialized all attributes for Base" << '\n';
     }
+    int base_pub_;
 
    protected:
-    int base_;
+    int base_pro_;
+
+   private:
+    int base_pri_;
 };  // class Base
 
-class Derived : public Base {
+class DerivedPublic : public Base {
    public:
-    Derived(double value)
-        : derived_{value} {
-        std::cout << "----- Derived ctor" << '\n';
-        std::cout << "base_: " << base_ << '\n';
-        std::cout << "derived_: " << derived_ << '\n';
+    DerivedPublic() : derived_{1} {}
+    void print_members() {
+        std::cout << "Printing from DerivedPublic" << '\n';
+        std::cout << base_pub_ << '\n';
+        std::cout << base_pro_ << '\n';
+        // std::cout << base_pri_ << '\n';
     }
 
    private:
-    double derived_;
-};  // class Derived
+    int derived_;
+};  // class DerivedPublic
+
+class DerivedProtected : protected Base {
+   public:
+    DerivedProtected() : derived_{2} {}
+    void print_members() {
+        std::cout << "Printing from DerivedProtected" << '\n';
+        std::cout << base_pub_ << '\n';
+        std::cout << base_pro_ << '\n';
+        // std::cout << base_pri_ << '\n';
+    }
+
+   private:
+    int derived_;
+};  // class DerivedProtected
+
+class DerivedPrivate : private Base {
+   public:
+    DerivedPrivate() : derived_{3} {}
+    void print_members() {
+        std::cout << "Printing from DerivedPrivate" << '\n';
+        std::cout << base_pub_ << '\n';
+        std::cout << base_pro_ << '\n';
+        // std::cout << base_pri_ << '\n';
+    }
+
+   private:
+    int derived_;
+};  // class DerivedPrivate
