+++ lecture8/include/person.hpp
@@ -1,4 +1,16 @@
 #pragma once
+
+/**
+ * @file person.hpp
+ * @author your name (you@domain.com)
+ * @brief
+ * @version 0.1
+ * @date 2024-10-14
+ *
+ * @copyright Copyright (c) 2024
+ *
+ */
+
 #include <cstddef>  // for offsetof
 #include <iostream>
 #include <memory>
@@ -12,11 +24,13 @@ struct Date {
     Date(unsigned short day,
          unsigned short month,
          unsigned short year) {
-            // validate first
         day_ = day;
         month_ = month;
         year_ = year;
     }
+
+    bool is_valid const(){}
+    bool is_leap_year const(){}
 };  // struct Date
 
 class Person {
@@ -24,9 +38,10 @@ class Person {
     Person() {}
     Person(const std::string& name, const std::string& email, const std::string& phone, unsigned short day, unsigned short month, unsigned short year) : name_{name}, email_{email}, phone_{phone}, date_of_birth_{std::make_unique<Date>(day, month, year)} {
     }
-
+    //---------- Mutators
+    //---------- Accessors
     std::string get_name() const { return name_; }
-
+    //---------- Utility
     void get_attribute_offsets() const {
         std::cout << "Offset of 'name_': " << offsetof(Person, name_) << '\n';
         std::cout << "Offset of 'email_': " << offsetof(Person, email_) << '\n';
